src/requirements.txt
fastapi
transformers
python-dotenv
uvicorn
torch
wandb
pydantic
streamlit